using quill first shot

+++ src/main/scala/me/herzrasen/stash/domain/User.scala
@@ -2,4 +2,4 @@ package me.herzrasen.stash.domain
 
 import me.herzrasen.stash.domain.Roles._
 
-case class User(id: Long, name: String, password: String, role: Role)
+case class User(id: Int, name: String, password: String, role: Role)
